+++ src/server.rs
@@ -23,7 +23,7 @@ use oinq::{
 use semver::{Version, VersionReq};
 
 #[cfg(feature = "server")]
-pub use self::handler::{Handler, handle};
+pub use self::handler::{Handler, handle, handle_authorized};
 #[cfg(feature = "server")]
 pub use self::stream::process_event_stream;
 #[cfg(feature = "server")]